Cambios de diseño y tests
