+++ .firebase/hosting.c3Jj.cache
@@ -1,7 +1,6 @@
 app.js,1631235348570,38ef38c97bc8995399bd473038a38ca1c52fe43638de70806e4ffb283518dba8
-index.html,1630645406917,56ba04cd468f4c57aec8d5a2da5bb2a7f50869ced8cc4803628c9f90441da5a9
+index.html,1631715772387,f4bb80d98b831e088c5c710cf2cae91e05eac8d0bf2ed25b7cbf66f43e1b8f22
 routes.js,1631235348651,1bd78fdc73889cadda407020ee83fcf95843c0f16f56427cea72ed20b26370a5
-style.css,1631631135446,70f6bb1d802434499285843da1e09586b2aad0d3f49e69ac0c54c2c10dfc2a13
 assets/back.png,1629080240071,4fa1a792b9cc91609c6d96b4c981614c537f8ba1970037c0b7512f54ea2553d7
 assets/close.png,1630091448159,3f56c4d21d147672a005fea056190e6418243b7989d9df652b71b66e68dacdf1
 assets/comment.png,1629833925694,f81685aaac6e030711ec4fb47827048d21e6519eccd1899ba8eb5e69219003ae
@@ -26,12 +25,23 @@ assets/search.png,1629833925708,252673a3b41666c3fd69623c859caf794f9098ae16b526ae
 assets/trash-can.png,1629833925710,6c4378f2a189f677e9b33c33667cbc05464c81a96995a5cbb0d19fb9398f6efd
 assets/triangle.png,1630075569024,0f8d4e1246b07204098945828940d2ae7570f7285c4fb11f64cd0a2d8ed323f6
 assets/user.png,1629833925711,a1c1a3a2754bfa32d88ff5d92d8f5082fc414014b0181bad736c3e4e79fe42e8
-components/getPosts.js,1631235348574,99fb82af4548b42d2eb40e6a31ec7e21e071a8941a19dbca56f0c8f594106119
-components/logIn.js,1631235348585,038124ce2cd90b83160e0e0706963a078cc2a938de31fdae6c676ea8e4817178
-components/post.js,1631235348592,63c2b51350b17e44b4fcdda866744ab47f0f1a5b253c2258aab9999c2e988525
-components/profile.js,1631631135444,ba5f3a344ee54066fb7a15d2e25b6fbbba9d93ffb5b67bc56ca09d2bf2c5c198
-components/signUp.js,1631235348624,5e908f6e4698482bac8b67219daad7055ca1086e57ecc7154f2b547000788e40
-components/updatePost.js,1631235348627,b1f05e4db7d1f188ca840b8b32eb717fc48f316c6ad237a54d0fabaec186fed3
+assets/readMe/Baja fidelidad.png,1631664204151,1ada88ea28340b69611a4d22714defd783a77564288db583a0241e0050fc3c9a
+assets/readMe/Bienvenida 1.png,1631665094317,f3b7c8fbe1486270f7be5fab1e8dde2c0e01ab1ce6130c1b89039cb2773791b9
+assets/readMe/Bienvenida 2.png,1631665102995,d58b54190ee313949cef0ff3531394d3e4e8f3fe119176f1bd96a0ac6b07bcc9
+assets/readMe/Bienvenida 3.png,1631665112190,4b55c4056acf7402258623faa1d5bf79229b8740333a9ca5b498697cafd6807c
+assets/readMe/Bienvenida 4.png,1631665119489,36c13611a309fb832201004171d3ba2b934c6642e276b78f7915fb801d01b11d
+assets/readMe/Bienvenida.png,1631663851160,c40329049a356a2f7485931ba0cf5a24547c13253579d9ecf03412dd7216791b
+assets/readMe/Crear Post.png,1631663883510,8480b4bc4f5ed224216cf136791a358bb4f3b46d84d15684641524218b83b0db
+assets/readMe/Crear Post1.png,1631665135152,7cb9208b60746f09593cfda7d5094c59bc9949102be638fe89744b5faf63926d
+assets/readMe/Home.png,1631663871570,0a31c70cf88f9b8118f828caf9769092f883a98615f7fd240aebb8b68a0af16e
+assets/readMe/Registro.png,1631663861495,cfb0fdbab971dec0899bb9161a3ebb8dd46303b25a03b740905e67ea93bd6626
+components/getPosts.js,1631661011372,7ac0a22b27d6cfd702903e72da7145e19add28690ad22a8437593b1126f2ad56
+components/home.js,1631716380339,f14558361404de3b7800fed514e7442c35ee8c18e774098a63d8ee099e9853ad
+components/logIn.js,1631661188871,a1938e935ab74cf6db2fce744bd62c8643ecac3c650f6bf0ec1a70f59280a448
+components/post.js,1631661391530,03536491825a8f97ad8a33e77e01cad1f9649abdecb03d9447c09de848c64e48
+components/profile.js,1631716395258,852a6233398f6f22157079117d8e1b0a645f423ce973b78abd60e0fd6b82fe3f
+components/signUp.js,1631661448006,acfacdc6dfa7f11ff55262caf4133d3db261d82a5abbf16548431ff79f4f1056
+components/updatePost.js,1631661503606,c50f6eef74fc048bea25fdb8f228c69f02a2350d56cda54a8da7207594caabbf
 lib/firebase.js,1631395237554,4110b0ae311ef0be1ec5a5fe391ea2dc9ee46a40676978d9e4b6e74f9f3766de
 lib/firebaseConfig.js,1629903422690,2775617a4c25e6f8af87833407451599e947df8447169915ad563ed74d45a52f
-components/home.js,1631642336303,14dde94459bf19202b16bf56dd39cf07357ef8c34f976e8b2777814bdfa8a956
+style.css,1631722556186,1ce4dc82f23c5a1efe6fe6657bb2db495b66b3a38fe8e0fb1a0699b9eb3ff5dd
