+++ test/logIn.spec.js
@@ -1,6 +1,8 @@
 /**
  * @jest-environment jsdom
  */
+/* eslint-disable quotes */
+//  import { onNavigate } from '../src/app.js';
 import './globals/firebase-test.js';
 import { logIn } from '../src/components/logIn.js';
